#1818, breakout subgraph generation into infer utilities
